+++ src/SimpleMessageFormatter.php
@@ -42,7 +42,7 @@ class SimpleMessageFormatter implements MessageFormatterInterface
 
             switch ($format) {
                 case 'plural':
-                    $options = $parts[2];
+                    $options = $parts[2] ?? '';
                     $replacements[$match] = self::pluralize($value, $options);
 
                     break;
@@ -62,24 +62,42 @@ class SimpleMessageFormatter implements MessageFormatterInterface
      */
     private static function pluralize($value, string $options): string
     {
+        if (!$options) {
+            throw new InvalidArgumentException('Missing plural keys: ' . self::formatList(self::PLURAL_KEYS) . '.');
+        }
+
         if (!is_int($value)) {
             throw new InvalidArgumentException('Only integer numbers are supported with plural format.');
         }
 
         preg_match_all('/([^{}\s]+)({(.*?)})/', $options, $pluralMatches);
 
-        foreach ($pluralMatches[1] as $match) {
+        $map = [];
+        foreach ($pluralMatches[1] as $index => $match) {
             if (!in_array($match, self::PLURAL_KEYS, true)) {
-                $keysStr = implode(', ', array_map(fn (string $value): string => '"' . $value . '"', self::PLURAL_KEYS));
+                $keysStr = self::formatList(self::PLURAL_KEYS);
 
                 throw new InvalidArgumentException("Invalid plural key - \"$match\". The valid keys are $keysStr.");
             }
+
+            $map[$match] = $pluralMatches[3][$index];
         }
 
-        $map = array_combine($pluralMatches[1], $pluralMatches[3]);
-        $formattedValue = $value . ' ';
-        $formattedValue .= $value === 1 ? $map[self::PLURAL_ONE] : $map[self::PLURAL_OTHER];
+        $diff = array_diff(self::PLURAL_KEYS, $pluralMatches[1]);
+        if ($diff !== []) {
+            throw new InvalidArgumentException('Missing plural keys: ' . self::formatList($diff) . '.');
+        }
 
-        return $formattedValue;
+        return $value === 1 ? $map[self::PLURAL_ONE] : $map[self::PLURAL_OTHER];
+    }
+
+    /**
+     * @param string[] $items
+     *
+     * @return string
+     */
+    private static function formatList(array $items): string
+    {
+        return implode(', ', array_map(fn (string $value): string => '"' . $value . '"', $items));
     }
 }